(FIX) make hook useLayoutHits web only as it breaks layout on mobile
